WIP commit for getting AOs involved with existing top containers
Due to ASpace not managing bidirectional relationships itself,
in order to collect existing AOs and sub_container data one has to lay their
grubby hands on the AOs involved with a top_container.  The cleanest way to
do this is via the search endpoint - this requires an up-to-date index, but works.